+++ changelog.md
@@ -4,3 +4,4 @@
   * Basic antigravity display published
 * 2020-11-11
   * Added bannerless variant
+  * Use databank value of targetAltitude if available to allow multiple screens to stay in sync